requirements.txt
Pillow==10.1.0
pycryptodome==3.19.0
python-dotenv==1.0.0
